The return purpose is to terminate the execution of the function, prevent continuing unrelated or unnecessary processing and prevent resolve(...); from executing.
Technically it is not needed here because a Promise can be resolved or rejected, exclusively and only once. The first Promise outcome wins and every subsequent result is ignored but it is a good clean practice to ensure that exactly one is called, and there is no further async/deferred processing.

source: http://stackoverflow.com/questions/32536049/do-i-need-to-return-after-early-resolve-reject